requirements.txt
# Web scraping
requests==2.31.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.2
playwright==1.49.0
fake-useragent==1.5.1
tenacity==8.3.0

# Data handling
pydantic==2.10.6
python-dotenv==1.0.1
PyYAML==6.0.2
orjson==3.10.15

# Azure Storage
azure-storage-blob==12.20.0
azure-identity==1.16.0

# Scheduling & CLI
schedule==1.2.2
click==8.1.7

# Logging
structlog==24.1.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-mock==3.14.0
responses==0.25.3

# Benchmarking
tabulate==0.9.0
